requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
PyMySQL==1.1.0
reportlab==4.0.4
python-dotenv==1.0.0
